% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textSpan.R
\name{textSpan}
\alias{textSpan}
\title{Textual Spanning Function}
\usage{
textSpan(A, alpha = 1)
}
\arguments{
\item{Input}{must be a square matrix of similarity scores.}
}
\description{
The function outputs a vector of cumulative textual spanning scores for each document.
}
\examples{
span.scores <- textSpan(sim0, alpha=1)
}
\references{
\url{https://journals.sagepub.com/doi/10.1177/2378023119827674}
}
